Simplify adding assets to loans
Previously one had to search for an asset first and could
then attach it to a loan. Because we only search with exact string
matching, there could only be one result.
Now the second step is gone and Assets are directly added.
If none is found, an error message is shown.