Add model factories and first feature tests
I decided not to use the database seeder for populating the
db during testing because in that case, the data population and testing
are too far apart.
With the current setup, I can seed the db with some fake data right
in the test class and use these data to check, if it gets properly
displayed.